Split nebula.go into arch specific files
